requirements.txt
Flask
mysql-connector-python
opencv-python
Pillow
gtts
pygame
numpy

dlib

jwt
pyjwt #pip install pyjwt